requirements.txt
Django==3.2.7
djangorestframework==3.11.1
